Langchain/requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
beautifulsoup4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
fastapi
uvicorn
langserve
sse_starlette 
duckduckgo-search
validators
youtube_transcript_api
unstructured
pytube
numexpr
huggingface_hub